backend-python/requirements.txt
fastapi
uvicorn
openai
python-dotenv
and hel center make that should as automation ai 